Added More Instructions; Fixed bugs; Added Debug Features
Added the BRK and SCN instructions. Added the Recursiondepthdata to
record the scope of running code. Changed to static linking. Had the exe
look for a .txt file instead of just getting the second parameter. Added
the .end header. Added more helpful error messages when problems are
related to scope.